+++ Changelogs/Changelog.md
@@ -18,3 +18,7 @@
 ### Fixed (73.4)
 - GW2 UI skin
 - Tabs order options
+
+### Fixed (73.5)
+- Fate of Nathria and Fate of Domination chains should now be correct
+- Most progress section in the achievement tooltip will no longer show for past achievements that are now unobtainable